+++ resources/views/layouts/includes/header.blade.php
@@ -25,12 +25,16 @@
                 </ul>
             </div>
             <button data-collapse-toggle="navbar-profile">
-                @if(Auth::user() && Auth::user()->profile_picture)
-                    <img class="w-10 h-10 rounded-full ring-2 ring-blue-500 mx-auto" src="{{ asset("storage/" . Auth::user()->profile_picture) }}" alt="Bordered avatar">
+                @if(Auth::user())
+                    @if(Auth::user()->profile_picture)
+                        <img class="object-cover w-10 h-10 rounded-full ring-2 ring-blue-500 mx-auto" src="{{ asset("storage/" . Auth::user()->profile_picture) }}" alt="Bordered avatar">
+                    @else
+                        <div class="p-1 ring-2 ring-blue-500 flex items-center justify-center w-10 h-10 overflow-hidden rounded-full bg-gray-600">
+                            <span class="font-medium text-gray-300">{{ \App\Services\ProfilePictureService::getProfilePictureInitials() }}</span>
+                        </div>
+                    @endif
                 @else
-                    <div class="p-1 ring-2 ring-blue-500 flex items-center justify-center w-10 h-10 overflow-hidden rounded-full bg-gray-600">
-                        <span class="font-medium text-gray-300">{{ \App\Services\ProfilePictureService::getProfilePictureInitials() }}</span>
-                    </div>
+                    <img class="object-cover w-10 h-10 rounded-full ring-2 ring-blue-500 mx-auto" src="{{ asset('images/idea-hub-logo-minimal.jpg') }}" alt="Bordered avatar">
                 @endif
             </button>
             <div id="navbar-profile" class="hidden w-5/12 md:w-4/12 lg:w-3/12 xl:w-2/12 absolute top-0 right-0 mt-20 flex-column justify-content-center align-items-center">
@@ -44,6 +48,7 @@
                             </form>
                         @else
                             <a href="{{ route('login') }}" class="block py-3 pl-3 pr-4 rounded hover:bg-gray-700 hover:text-blue-500 @if(request()->routeIs('login')) text-blue-500 @endif" aria-current="page">Log In</a>
+                            <a href="{{ route('register') }}" class="block py-3 pl-3 pr-4 rounded hover:bg-gray-700 hover:text-blue-500 @if(request()->routeIs('register')) text-blue-500 @endif" aria-current="page">Sign Up</a>
                         @endif
                     </li>
                 </ul>