+++ tests/Feature/Profile/UpdateProfilePictureTest.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Profile;
+namespace Tests\Feature\Profile;
 
+use App\Models\DefaultProfilePicture;
 use App\Models\User;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
@@ -11,15 +12,15 @@ class UpdateProfilePictureTest extends TestCase
 {
     public function test_profile_picture_is_saved_when_updating_profile_picture()
     {
+        Storage::fake();
+
         $user = User::factory()->create();
 
         // Log in as user
         $this->actingAs($user);
 
-        Storage::fake('public');
-
         // Update profile picture
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => UploadedFile::fake()->image('profile_picture.jpg', 100, 100)->size(100),
         ]);
 
@@ -46,7 +47,7 @@ class UpdateProfilePictureTest extends TestCase
         $this->actingAs($user);
 
         // Attempt to update profile picture without providing a profile picture
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => '',
         ]);
 
@@ -63,7 +64,7 @@ class UpdateProfilePictureTest extends TestCase
         $this->actingAs($user);
 
         // Attempt to upload profile picture that is 6MB
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => UploadedFile::fake()->image('profile_picture.jpg')->size(6000),
         ]);
 
@@ -80,7 +81,7 @@ class UpdateProfilePictureTest extends TestCase
         $this->actingAs($user);
 
         // Attempt to upload profile picture that is not an image
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => UploadedFile::fake()->create('profile_picture.pdf'),
         ]);
 
@@ -97,7 +98,7 @@ class UpdateProfilePictureTest extends TestCase
         $this->actingAs($user);
 
         // Attempt to upload profile picture with extension of bmp
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => UploadedFile::fake()->image('profile_picture.bmp'),
         ]);
 
@@ -106,7 +107,7 @@ class UpdateProfilePictureTest extends TestCase
         ]);
     }
 
-    public function test_profile_picture_resolution_must_be_800x800_or_less()
+    public function test_profile_picture_resolution_must_be_4000x4000_or_less()
     {
         $user = User::factory()->create();
 
@@ -114,17 +115,19 @@ class UpdateProfilePictureTest extends TestCase
         $this->actingAs($user);
 
         // Attempt to upload profile picture with resolution of 801x801
-        $response = $this->patch('/profile/profile-picture', [
-            'profile_picture' => UploadedFile::fake()->image('profile_picture.jpg', 801, 801),
+        $response = $this->patch('/profile-picture', [
+            'profile_picture' => UploadedFile::fake()->image('profile_picture.jpg', 4001, 4001),
         ]);
 
         $response->assertSessionHasErrors([
-            'profile_picture' => 'Profile picture must be 800x800 or less.',
+            'profile_picture' => 'Profile picture must be 4000x4000 or less.',
         ]);
     }
 
     public function test_old_uploaded_profile_picture_is_deleted_when_new_one_is_uploaded()
     {
+        Storage::fake();
+
         $user = User::factory()->create([
             'profile_picture' => 'profile_picture.jpg',
         ]);
@@ -133,11 +136,10 @@ class UpdateProfilePictureTest extends TestCase
         $this->actingAs($user);
 
         // Store fake profile picture
-        Storage::fake('public');
         UploadedFile::fake()->image('profile_picture.jpg', 100, 100)->size(100)->store('images/users/profile_pictures');
 
         // Update new profile picture
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => UploadedFile::fake()->image('new_profile_picture.jpg', 100, 100)->size(100),
         ]);
 
@@ -147,11 +149,13 @@ class UpdateProfilePictureTest extends TestCase
         ]);
 
         // Assert old profile picture is deleted
-        Storage::disk('public')->assertMissing('images/users/profile_pictures/profile_picture.jpg');
+        Storage::assertMissing('images/users/profile_pictures/profile_picture.jpg');
     }
 
     public function test_profile_picture_is_updated()
     {
+        Storage::fake();
+
         $user = User::factory()->create([
             'profile_picture' => 'profile_picture.jpg',
         ]);
@@ -159,11 +163,10 @@ class UpdateProfilePictureTest extends TestCase
         // Log in as user
         $this->actingAs($user);
 
-        // Store fake profile picture
-        Storage::fake('public');
-        UploadedFile::fake()->image('profile_picture.jpg', 100, 100)->size(100)->store('images/users/profile_pictures');
+        // Store fake profile picture in profile pictures directory
+        UploadedFile::fake()->image('profile_picture.jpg', 100, 100)->size(100)->store(config('filesystems.profile_pictures_path'));
 
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => UploadedFile::fake()->image('new_profile_picture.jpg', 100, 100)->size(100),
         ]);
 
@@ -185,10 +188,32 @@ class UpdateProfilePictureTest extends TestCase
     public function test_cannot_update_profile_picture_if_not_authenticated()
     {
         // Attempt to store fake profile picture while unauthenticated
-        $response = $this->patch('/profile/profile-picture', [
+        $response = $this->patch('/profile-picture', [
             'profile_picture' => UploadedFile::fake()->image('profile_picture.jpg', 100, 100)->size(100),
         ]);
 
         $response->assertRedirect('/login');
+    }
+
+    public function test_default_profile_picture_is_not_deleted_when_updating_profile_picture()
+    {
+        Storage::fake();
+
+        $defaultProfilePicture = DefaultProfilePicture::factory()->create();
+
+        // Give user a default uploaded profile picture
+        $user = User::factory()->create([
+            'profile_picture' => $defaultProfilePicture->path,
+        ]);
+
+        // Log in as user
+        $this->actingAs($user);
+
+        $this->patch('/profile-picture', [
+            'profile_picture' => UploadedFile::fake()->image('new_profile_picture.jpg', 100, 100)->size(100),
+        ]);
+
+        // assert default profile picture was not deleted
+        Storage::assertExists($defaultProfilePicture->path);
     }
 }